refactor: separate artifact entities and data access
